remove old tech test "Tests" Page; rename TestListPage to TestsPage; fix spacing for show configuration methods toggle
