+++ report-ng/app-react/src/pages/TestsPage.tsx
@@ -1,70 +1,86 @@
-import {useReportData} from "../provider/DataProvider.tsx";
 import Box from "@mui/material/Box";
-import {Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Typography} from "@mui/material";
-import {ResultStatusType} from "../model/report-model/framework_pb.ts";
+import {Grid, Switch, Typography} from "@mui/material";
+import StatusSelectInput from "../widgets/StatusSelectInput";
+import MultiSelectInput from "../widgets/MultiSelectInput";
+import {useTestListFilters} from "../hooks/useTestListFilters";
+import SelectedFiltersChips from "../components/SelectedFilterChips";
+import TestList from "../components/TestList";
+import SearchInput from "../widgets/SearchInput";
+import {useState} from "react";
+import Stack from "@mui/material/Stack";
 
-function createRow(
-    key: string,
-    value: number
-) {
-    return {key, value};
-}
+const TestsPage = () => {
+    const {
+        statusMenuItems,
+        classMenuItems,
+        filters,
+        setFilter,
+        configurationMethodsChecked,
+        handleConfigurationMethodsChecked,
+        handleDelete,
+        clearAll,
+    } = useTestListFilters();
 
-const Tests = () => {
+    const selectedStatuses = filters.status ?? [];
+    const selectedClasses = filters.class ?? [];
 
-    const {executionMngr, isLoading, error} = useReportData();
-
-    if (isLoading) return <div>Lade Konfiguration...</div>;
-    if (error) return <div>Fehler: {error.message}</div>;
-    if (!executionMngr) return null;
-
-    const execStatistics = executionMngr.getExecutionStatistics();
-
-    const rows = [
-        createRow('Number of tests', execStatistics.overallTestCases),
-        createRow('Overall passed', execStatistics.overallPassed),
-        createRow('Overall failed', execStatistics.overallFailed),
-        createRow('Passed', execStatistics.getStatusCount(ResultStatusType.PASSED)),
-        createRow('Repaired', execStatistics.getStatusCount(ResultStatusType.REPAIRED)),
-        createRow('Recovered', execStatistics.getStatusCount(ResultStatusType.PASSED_RETRY)),
-        createRow('Failed', execStatistics.getStatusCount(ResultStatusType.FAILED)),
-        createRow('Expected failed', execStatistics.getStatusCount(ResultStatusType.FAILED_EXPECTED)),
-        createRow('Failed retry', execStatistics.getStatusCount(ResultStatusType.FAILED_RETRIED)),
-        createRow('Skipped', execStatistics.getStatusCount(ResultStatusType.SKIPPED)),
-    ];
+    // live search text for highlighting while typing (not yet confirmed)
+    const [searchText, setSearchText] = useState("");
 
     return (
-        <>
-            <Box>
-                <Typography variant="h6">Test view</Typography>
-                <Typography>
-                    {executionMngr.getExecutionAggregate().executionContext?.runConfig?.reportName}
-                </Typography>
-                <TableContainer component={Paper}>
-                    <Table sx={{minWidth: 200}} aria-label="simple table">
-                        <TableHead>
-                            <TableRow>
-                                <TableCell>Statistic</TableCell>
-                                <TableCell align="center">Value</TableCell>
-                            </TableRow>
-                        </TableHead>
-                        <TableBody>
-                            {rows.map((row) => (
-                                <TableRow
-                                    key={row.key}
-                                    sx={{'&:last-child td, &:last-child th': {border: 0}}}
-                                >
-                                    <TableCell component="th" scope="row">
-                                        {row.key}
-                                    </TableCell>
-                                    <TableCell align="right">{row.value}</TableCell>
-                                </TableRow>
-                            ))}
-                        </TableBody>
-                    </Table>
-                </TableContainer>
-            </Box>
-        </>
+        <Box
+            sx={{width: '100%', maxWidth: {sm: '100%', md: '1700px'}}}
+        >
+            <Grid
+                container
+                spacing={3}
+                columns={12}
+            >
+                <Grid size={2}>
+                    <StatusSelectInput label="Status" selectedStatuses={selectedStatuses}
+                                       onChange={(newStatuses) => setFilter("status", newStatuses)}
+                                       menuItems={statusMenuItems}/>
+                </Grid>
+                <Grid size={3}>
+                    <MultiSelectInput label="Class"
+                                      values={selectedClasses}
+                                      onChange={(newClasses) => setFilter("class", newClasses)}
+                                      menuItems={classMenuItems}
+                                      renderValue={(selected: string[]) => {
+                                          if (!selected?.length) return "";
+                                          if (selected.length === 1) return "1 class selected";
+                                          return `${selected.length} classes selected`;
+                                      }}/>
+                </Grid>
+                <Grid size={5}>
+                    <SearchInput
+                        currentTexts={filters.customText ?? []}
+                        onConfirm={(newTexts) => {
+                            setFilter("customText", newTexts);
+                            setSearchText("");
+                        }}
+                        onSearchTextChange={setSearchText}
+                    />
+                </Grid>
+                <Grid size={2} sx={{alignContent: "center"}}>
+                    <Stack direction="row" sx={{alignItems: "center"}}>
+                        <Switch checked={configurationMethodsChecked}
+                                onChange={handleConfigurationMethodsChecked}/>
+                        <Typography variant="body2">Show configuration methods</Typography>
+                    </Stack>
+                </Grid>
+                <Grid size={12} minHeight={36}>
+                    <SelectedFiltersChips selectedFilters={filters}
+                                          handleDelete={handleDelete}
+                                          handleClearAllClick={clearAll}/>
+
+                </Grid>
+                <Grid size={12} >
+                    <TestList filters={filters} searchText={searchText} showConfigurationMethods={configurationMethodsChecked}/>
+                </Grid>
+            </Grid>
+        </Box>
+
     );
 };
-export default Tests;
+export default TestsPage;
